Fix grep-hist -first.
grep-hist -first should find the first revision that doesn't match the
pattern, then show the lastrev of the next revision, which would be
the revision that added a line that matched the pattern.

Before, it was incorrectly showing the lastrev of the next revision
that touched a line that matched the pattern.