More lazy evaluation: Change read_input so that it represents tables as generators of rows.
This means we can do a better job of separating input (reading files)
and output (printing the table) -- in particular, "immediate" mode can
be safely moved to pretty_print, and the idea of nopretty (which
affects how data is output) can be more cleanly separated from raw
(which affects the munging of fields on the input side).  This
separation is more important with sql mode, where there is some
processing going on between the input and output.