game_spaces.py: fix bug in make_transition_matrix
spotter: podolak

The problem was that I had to flip the probabilities when I put them
into the transition matrix, based on how I was building the matrix.  I
defined the output array such that after multiplying by the transition
matrix N times, arr[-1] = the probability of landing on space N, and
arr[-i] = probability of landing on space N-i+1 -- i.e., it's sort of
backwards, the last element in the array is the "most recent space".
So the transition matrix is mostly 1s just above the diagonal (that's
what moves arr[i-1] = arr[i]) except the last row of the matrix should
be the probabilities of each step size: if the last row is [x1, x2,
x3, x4, ..., x12], then x12 should be the probability of moving one
space and x11 should be the probability of moving 2 spaces, etc.  I
had it flipped, I had x1 as the probability of moving one space and
x12 the probability of moving 12 spaces, etc.  It didn't matter for
the case where the probability of all spaces is the same, but it
messed things up for dice where there is no way of moving just one
space.