Add forgiving mode to ProtoList and ProtoDict
In forgiving mode, if you ask for a field that doesn't exist, it doesn't throw an exception, it just gives you an empty ProtoDict or ProtoList, (which is itself forgiving, so you can just keep asking for fields that don't exist).  This is good when you have a list of protos that you want to inspect, and the field you want to see is deep in the protos' internals, and it might exist on some of the protos but not others.